refactor: update delete offer endpoint logic and adjust tests
- Refactored DELETE endpoint '/offers/{IDOferta}' to change the deletion logic
- Updated unit tests to reflect the new deletion logic

+++ src/offer/offerService.js
@@ -1,5 +1,6 @@
 import { FailedToDeleteError } from '../errors/ErrorTypes/failedToDeleteError.js';
-import { FAILED_DELETING, OFFERS } from '../utils/textConstants.js';
+import { InvalidDataError } from '../errors/ErrorTypes/invalidDataError.js';
+import { FAILED_DELETING, INVALID_DATA, OFFERS } from '../utils/textConstants.js';
 
 export class OfferService {
   constructor ({ offerRepository }) {
@@ -12,10 +13,12 @@ export class OfferService {
   };
 
   deleteOffer = async (id) => {
+    if (!id) {
+      throw new InvalidDataError(INVALID_DATA, OFFERS);
+    }
     const isDelete = await this.offerRepository.deleteOffer(id);
-    if (isDelete.success !== true) {
+    if (isDelete !== true) {
       throw new FailedToDeleteError(FAILED_DELETING, OFFERS);
     }
-    return isDelete;
   };
 }
